refactor: rename db to dao
