+++ slipp/src/test/java/slipp/dao/UserDaoTest.java
@@ -1,4 +1,4 @@
-package slipp.db;
+package slipp.dao;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
